refactor: nest private parts, rename TripsTableHeaderMenu
A component only its parent can ever render now lives in the parent's
directory rather than at the domain root:

- TripsMobileListItem -> TripsMobileList/
- MileageDeductionChartTooltip -> MileageDeductionChart/

That also makes MileageDeductionChartDataPoint private to the chart
again, instead of a shared type sitting inside one of two consumers.

TripsHeaderMenu (was TripsTableHeaderMenu) stays at the domain root: it
is rendered by both TripsTableHeader and TripsMobileHeader, so nesting
it would make one of them reach into the other's private directory. The
old name implied the table header owned it.
