+++ src/components/features/mileage/TripsTableHeader/TripsTableHeader.tsx
@@ -10,7 +10,7 @@ import { Button } from '@ui/Button/Button'
 import { HStack } from '@ui/Stack/Stack'
 import { DataTableHeader } from '@blocks/DataTable/DataTableHeader'
 import { TripPurposeToggle } from '@features/mileage/TripPurposeToggle/TripPurposeToggle'
-import { TripsTableHeaderMenu } from '@features/mileage/TripsTableHeaderMenu/TripsTableHeaderMenu'
+import { TripsHeaderMenu } from '@features/mileage/TripsHeaderMenu/TripsHeaderMenu'
 import { VehicleSelector } from '@features/mileage/VehicleSelector/VehicleSelector'
 
 import './tripsTableHeader.scss'
@@ -55,7 +55,7 @@ export const TripsTableHeader = ({ onRecordTrip }: TripsTableHeaderProps) => {
         {t('trips:action.record_trip', 'Record Trip')}
         <Plus size={16} />
       </Button>
-      <TripsTableHeaderMenu />
+      <TripsHeaderMenu />
     </HStack>
   ), [t, onRecordTrip])
